move borrowing and decommissioning into tools app
