+++ toolhub/urls.py
@@ -14,8 +14,6 @@ urlpatterns = [
     path("", include("toolhub_auth.urls"), name="auth"),
     path("admin/", admin.site.urls),
     path("tools/", include("tools.urls"), name="tools"),
-    path("borrowing/", include("borrowing.urls"), name="borrowing"),
-    path("decommissioning/", include("decommissioning.urls"), name="decomissioning"),
     # endpoints
     path("markdownx/", include("markdownx.urls")),
     # shortcut urls
