+++ toolhub_auth/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = "toolhub_auth.apps.ToolhubAuthConfig"
